fix: support log error in list command
