+++ packages/core/src/utils/index.ts
@@ -2,3 +2,4 @@ export * from './helper';
 export * from './logger';
 export * from './testFiles';
 export * from './constants';
+export * from './error';
